src/asl-engine/requirements.txt
ipykernel==6.16.0
ipympl==0.9.2
ipython==8.5.0
ipywidgets==8.0.2
matplotlib==3.6.1
matplotlib-inline==0.1.6
mediapipe-silicon==0.8.11
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.5.0
path==16.5.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
plyfile==0.7.4
python-dotenv==0.21.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sklearn==0.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
